Kill process when forfeit-time
So now when a player takes too much time to move, they forfiet without having to wait for them to actually move (which will never happen when the player has entered an infinite loop).  However, the standard-output from that player gets lost.  It would be nice if we could avoid that.